Added a BaseLoader template loader for code factoring and some tmeplate loading optimization.
Template loading optimization is based on the new settings FLAVOURS_NOMOBILE_TEMPLATE which tells us
whether or not flavouring a template when a non-mobile template is found. Also, when an already
flavoured template is found, we skip it.
The non-mobile flavour is defined by the new DEFAULT_NOMOBILE_FLAVOUR settings.